+++ lib/experiment_information.dart
@@ -7,7 +7,7 @@ class ExperimentInformation {
   final String currentVariant;
   final Map<String, Map<String, dynamic>> variants;
 
-  static ExperimentInformation fromJson(dynamic json) {
+  static ExperimentInformation fromJson(Map json) {
     Map<String, Map<String, dynamic>> variantsMap = {};
     Map<Object?, Object?> variants = json['variants'] ?? {};
     for (var item in variants.keys)
@@ -16,7 +16,7 @@ class ExperimentInformation {
       Map<Object?, Object?> values = variants[item] as Map<Object?, Object?>;
       for (var key in values.keys)
       {
-        valueMap[key.toString()] = variants[key];
+        valueMap[key.toString()] = values[key];
       }
       variantsMap[item.toString()] = valueMap;
     }