+++ lib/remote_config_internal.dart
@@ -105,6 +105,50 @@ class RemoteConfigInternal implements RemoteConfig {
   }
 
   @override
+  Future<void> testingEnrollIntoABExperiment(String experimentID) async {
+    final Map<String, ExperimentInformation> experimentsInfoMap = await testingGetAllExperimentInfo();
+    final ExperimentInformation? experimentInformation = experimentsInfoMap[experimentID];
+    if (experimentInformation == null) {
+      Countly.log("testingExitABExperiment, No experiment information found against experiment Id: '$experimentID'", logLevel: LogLevel.WARNING);
+      return;
+    }
+
+    if (experimentInformation.variants.isEmpty) {
+      Countly.log("testingExitABExperiment, No variants found in experiment information against experiment Id: '$experimentID'", logLevel: LogLevel.WARNING);
+      return;
+    }
+
+    if (experimentInformation.variants.entries.first.value.keys.isEmpty) {
+      Countly.log("testingExitABExperiment, No values found against in variants for experiment information against experiment Id: '$experimentID'", logLevel: LogLevel.WARNING);
+      return;
+    }
+
+    await enrollIntoABTestsForKeys(experimentInformation.variants.entries.first.value.keys.toList());
+  }
+
+  @override
+  Future<void> testingExitABExperiment(String experimentID) async {
+    final Map<String, ExperimentInformation> experimentsInfoMap = await testingGetAllExperimentInfo();
+    final ExperimentInformation? experimentInformation = experimentsInfoMap[experimentID];
+    if (experimentInformation == null) {
+      Countly.log("testingExitABExperiment, No experiment information found against experiment Id: '$experimentID'", logLevel: LogLevel.WARNING);
+      return;
+    }
+
+    if (experimentInformation.variants.isEmpty) {
+      Countly.log("testingExitABExperiment, No variants found in experiment information against experiment Id: '$experimentID'", logLevel: LogLevel.WARNING);
+      return;
+    }
+
+    if (experimentInformation.variants.entries.first.value.keys.isEmpty) {
+      Countly.log("testingExitABExperiment, No values found against in variants for experiment information against experiment Id: '$experimentID'", logLevel: LogLevel.WARNING);
+      return;
+    }
+
+    await exitABTestsForKeys(experimentInformation.variants.entries.first.value.keys.toList());
+  }
+
+  @override
   Future<void> enrollIntoABTestsForKeys(List<String> keys) async {
     if (!_countlyState.isInitialized) {
       Countly.log('"initWithConfig" must be called before "remoteConfigEnrollIntoABTestsForKeys"', logLevel: LogLevel.ERROR);
@@ -216,7 +260,7 @@ class RemoteConfigInternal implements RemoteConfig {
   }
 
   @override
-  Future<Map<String, RCData>> getAllValuesAndEnroll()  async {
+  Future<Map<String, RCData>> getAllValuesAndEnroll() async {
     if (!_countlyState.isInitialized) {
       Countly.log('"initWithConfig" must be called before "getAllValuesAndEnroll"', logLevel: LogLevel.ERROR);
       return {};
@@ -333,14 +377,14 @@ class RemoteConfigInternal implements RemoteConfig {
     }
     Countly.log(key.toString());
 
-    List<String> args = [];
+    final List<String> args = [];
     args.add(key);
 
     List<dynamic>? returnValue = await _countlyState.channel.invokeMethod('remoteConfigTestingGetVariantsForKey', <String, dynamic>{'data': json.encode(args)});
 
     returnValue ??= [];
 
-    List<String> variant = List<String>.from(returnValue);
+    final List<String> variant = List<String>.from(returnValue);
 
     return variant;
   }
@@ -352,25 +396,24 @@ class RemoteConfigInternal implements RemoteConfig {
       return;
     }
     Countly.log('Calling "testingDownloadExperimentInformation"');
-    int requestID = _wrapVariantCallback(rcVariantCallback);
+    final int requestID = _wrapVariantCallback(rcVariantCallback);
 
-    List<dynamic> args = [];
+    final List<dynamic> args = [];
     args.add(requestID);
 
     return await _countlyState.channel.invokeMethod('testingDownloadExperimentInformation', <String, dynamic>{'data': json.encode(args)});
   }
 
   @override
-  Future<Map<String, ExperimentInformation>> testingGetAllExperimentInfo() async
-  {
+  Future<Map<String, ExperimentInformation>> testingGetAllExperimentInfo() async {
     if (!_countlyState.isInitialized) {
       Countly.log('"initWithConfig" must be called before "testingGetAllExperimentInfo"', logLevel: LogLevel.ERROR);
       return {};
     }
 
     final List<dynamic> experimentsInfo = await _countlyState.channel.invokeMethod('testingGetAllExperimentInfo');
-    List<ExperimentInformation> experimentsInfoList = experimentsInfo.map(ExperimentInformation.fromJson).toList();
-    Map<String, ExperimentInformation> experimentsInfoMap = Map.fromIterable(experimentsInfoList, key: (e) => e.experimentID, value: (e) => e);
+    final List<ExperimentInformation> experimentsInfoList = experimentsInfo.map((e) => ExperimentInformation.fromJson(e)).toList();
+    final Map<String, ExperimentInformation> experimentsInfoMap = {for (final e in experimentsInfoList) e.experimentID: e};
     return experimentsInfoMap;
   }