+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 ## 23.8.4
+* Added a call to enroll users to A/B experiment with experiment ID : `testingEnrollIntoABExperiment:`
+* Added a call to exit users from A/B experiment with experiment ID : `testingExitABExperiment:`
+
 * Fixed the exit AB request failure issue on iOS.
-* Updated underlying Android SDK version to 23.8.3
+* Fixed 'null' value issue for experiment info variants.
+* Fixed 'null' pointer error thrown when 'testingGetVariantsForKey' method is called with a key that does not exist.
+
+* Updated underlying Android SDK version to 23.8.4
 * Underlying iOS SDK version is 23.8.3
 
 ## 23.8.4-np
+* Added a call to enroll users to A/B experiment with experiment ID : `testingEnrollIntoABExperiment:`
+* Added a call to exit users from A/B experiment with experiment ID : `testingExitABExperiment:`
+
 * Fixed the exit AB request failure issue on iOS.
-* Updated underlying Android SDK version to 23.8.3
+* Fixed 'null' value issue for experiment info variants.
+* Fixed 'null' pointer error thrown when 'testingGetVariantsForKey' method is called with a key that does not exist.
+
+* Updated underlying Android SDK version to 23.8.4
 * Underlying iOS SDK version is 23.8.3
 
 ## 23.8.3